# doc/API_Specification_OAS3.yaml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: "Organz API"
  description: |
    An API used to access the Human Harvest Inc server.
    Versioning: Adding new endpoints, adding parameters or key value pairs are all minor changes, as they are still valid with a past client.
      Major changes include: Modifying the data types of any parameters or values, removing key value pairs, removing endpoints. These CANNOT be pushed without maintaining a legacy version and upgrading the system to a new version.
...
paths:
...
  /clinicians:
    get:
      tags:
        - Clinician
      security:
        - ClinicianAccess: []
        - AdministratorAccess: []
      summary: "Returns all clinicians from the system"
      parameters:
        - $ref: '#/components/parameters/q'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/count'
        - $ref: '#/components/parameters/region'
      responses:
        200:
          description: A list of clinicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/clinicianOverview'

        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        500:
          $ref: '#/components/responses/500InternalServerError'
    post:
      tags:
        - Clinician
      security:
        - AdministratorAccess: []
      summary: Create a new clinician
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/clinicianCreateDetails'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/clinicianDetails'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        500:
          $ref: '#/components/responses/500InternalServerError'
  '/clinicians/{staffId}':
    parameters:
    - in: path
      name: staffId
      description: Clinician Staff ID
      required: true
      schema:
        type: integer
        minimum: 0
    get:
      tags:
        - Clinician
      security:
        - ClinicianAccess: []
        - AdministratorAccess: []
      summary: Get detailed clinician information
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/clinicianDetails'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalServerError'
    patch:
      tags:
        - Clinician
      security:
        - ClinicianAccess: []
        - AdministratorAccess: []
      summary: Change some information about a clinician
      description: Only those properties included in the request body will be changed.
      parameters:
        - $ref: '#/components/parameters/etag'

      requestBody:
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/clinicianModifiableDetails'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/clinicianDetails'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        404:
          $ref: '#/components/responses/404NotFound'
        412:
          $ref: '#/components/responses/412PreconditionFailed'
        428:
          $ref: '#/components/responses/428PreconditionRequired'
        500:
          $ref: '#/components/responses/500InternalServerError'
    delete:
      tags:
        - Clinician
      security:
        - ClinicianAccess: []
        - AdministratorAccess: []
      summary: Delete a clinician
      parameters:
        - $ref: '#/components/parameters/etag'

      responses:
        201:
          description: Deleted
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        404:
          $ref: '#/components/responses/404NotFound'
        412:
          $ref: '#/components/responses/412PreconditionFailed'
        428:
          $ref: '#/components/responses/428PreconditionRequired'
        500:
          $ref: '#/components/responses/500InternalServerError'




  ### Administrator
...
components:
  parameters:
    q:
      name: q
      in: query
      description: "The search string"
      schema:
        type: string
    offset:
      name: offset
      in: query
      description: "The offset from the start of the result set, used for pagination"
      schema:
        type: integer
    count:
      name: count
      in: query
      description: "How many results to return, used for pagination"
      schema:
        type: integer
...
    region:
      name: region
      in: query
      description: "The region to filter, may be comma seperated for multiple"
      schema:
        $ref: '#/components/schemas/regionEnum'
...
    etag:
      name: etag
      in: header
      description: A hashed value of the object used for optimistic concurrency control
      required: true
      schema:
        $ref: '#/components/schemas/etag'
...
  schemas:
...
    clinicianOverview:
      type: object
      properties:
        staffId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        region:
          $ref: '#/components/schemas/regionEnum'
    clinicianDetails:
      type: object
      properties:
        staffId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        workAddress:
          type: string
        region:
          $ref: '#/components/schemas/regionEnum'
        createdOn:
          $ref: '#/components/schemas/dateTime'
        modifiedOn:
          $ref: '#/components/schemas/dateTime'
    clinicianModifiableDetails:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        workAddress:
          type: string
        password:
          type: string
        region:
          $ref: '#/components/schemas/regionEnum'
    clinicianCreateDetails:
      type: object
      properties:
        staffId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        workAddress:
          type: string
        password:
          type: string
        region:
          $ref: '#/components/schemas/regionEnum'


    ### Administrator
...
    dateTime:
      type: string
      example: "2018-07-19T14:56:01.562"
...
    regionEnum:
      type: string
      enum: [NORTHLAND, AUCKLAND, WAIKATO, BAY_OF_PLENTY, GISBORNE, HAWKES_BAY, TARANAKI, MANAWATU_WANGANUI, WELLINGTON, TASMAN, NELSON, MARLBOROUGH, WEST_COAST, CANTERBURY, OTAGO, SOUTHLAND, UNSPECIFIED]
...
    etag:
      type: string
      description: A hashed value of the object used for optimistic concurrency control
  responses:
    400BadRequest:
      description: Bad request
    401Unauthorized:
      description: Access token is missing or invalid
    403Forbidden:
      description: You do not have permission to perform that action
    404NotFound:
      description: The specified resource was not found
    412PreconditionFailed:
      description: That resource has already been modified or the ETag header is invalid
    428PreconditionRequired:
      description: ETag header was missing and is required to modify a resource
    500InternalServerError:
      description: Internal server error
  securitySchemes:
          
    ClientAccess:
      description: A client may only access their own data, they cannot access other users. They may only modify their personal information and organ donation status. They can view all their other statuses.
      type: http
      scheme: bearer
      bearerFormat: JWT
      
    ClinicianAccess:
      description: Clincians can access and modify and client, and may view any clinician, but only modify their own information.
      type: http
      scheme: bearer
      bearerFormat: JWT
      
    AdministratorAccess:
      description: Administrators can access and modify all user types.
      type: http
      scheme: bearer
      bearerFormat: JWT